+++ tf_card.c
@@ -99,9 +99,12 @@ static pio_spi_inst_t _pio_spi = {
 static PIO _pio = SPI_PIO_DEFAULT_PIO;
 static uint _sm = SPI_PIO_DEFAULT_SM;
 static io_rw_32* _reg_clkdiv = NULL;
-static io_rw_32  _pio_clkdiv_slow = 255 << 16;
-static io_rw_32  _pio_clkdiv_fast = 255 << 16;
+static io_rw_32  _pio_clkdiv_slow = 4096 << 16;
+static io_rw_32  _pio_clkdiv_fast =  256 << 16;
+#define PIO_CLKDIV_LIMIT (0x00018000)  // fractional div x1.5 (6 system clock syscles per 1 SCK cycle)
 
+static uint _clk_slow_freq = CLK_SLOW_DEFAULT;
+static uint _clk_fast_freq = CLK_FAST_DEFAULT;
 
 static inline uint32_t _millis(void)
 {
@@ -172,7 +175,7 @@ static void pico_fatfs_init_spi_pio(void)
     uint offset = pio_add_program(_pio_spi.pio, &spi_cpha0_program);
     pio_spi_init(_pio_spi.pio, _pio_spi.sm, offset,
         8,       // 8 bits per SPI frame
-        (float) f_clk_sys / (_config.clk_slow / KHZ) / 4,
+        (float) f_clk_sys / (_config.clk_slow / KHZ) / 4,  // 4 PIO input clock cycles to generate one SPI clock cycle
         false,   // CPHA = 0
         false,   // CPOL = 0
         _config.pin_sck,
@@ -183,6 +186,13 @@ static void pico_fatfs_init_spi_pio(void)
     _reg_clkdiv      = &(_pio_spi.pio->sm[_pio_spi.sm].clkdiv);
     _pio_clkdiv_slow = *_reg_clkdiv;
     _pio_clkdiv_fast = (io_rw_32) ((float) _pio_clkdiv_slow * _config.clk_slow / _config.clk_fast);
+    _pio_clkdiv_fast &= 0xffffff00;
+    if (_pio_clkdiv_fast < PIO_CLKDIV_LIMIT) {
+        _pio_clkdiv_fast = PIO_CLKDIV_LIMIT;
+    }
+
+    _clk_fast_freq = (uint64_t) f_clk_sys * 1000 * 256 / (_pio_clkdiv_fast / 256)  / 4;
+    _clk_slow_freq = (uint64_t) f_clk_sys * 1000 * 256 / (_pio_clkdiv_slow / 256)  / 4;
 }
 
 /* Initialize SPI */
@@ -242,6 +252,11 @@ void pico_fatfs_init_spi(void)
         SPI_CPHA_0, /* cpha */
         SPI_MSB_FIRST /* order */
     );
+
+    FCLK_FAST();
+    _clk_fast_freq = spi_get_baudrate(_config.spi_inst);
+    FCLK_SLOW();
+    _clk_slow_freq = spi_get_baudrate(_config.spi_inst);
 }
 
 /* Exchange a byte */
@@ -701,6 +716,9 @@ bool pico_fatfs_set_config(pico_fatfs_spi_config_t* config)
 {
     _config = *config;
 
+    _clk_slow_freq = CLK_SLOW_DEFAULT;
+    _clk_fast_freq = CLK_FAST_DEFAULT;
+
     if (_config.spi_inst == NULL) {
         return false;
     } else if (_config.spi_inst != spi0 && _config.spi_inst != spi1) {
@@ -747,5 +765,26 @@ void pico_fatfs_config_spi_pio(PIO pio, uint sm)
 
 int pico_fatfs_reboot_spi(void)
 {
+    // Send 32 cycles of CS low
+    FCLK_SLOW();
+    CS_HIGH();
+    sleep_ms(10);
+    CS_LOW();
+    for (int i = 0; i < 4; i++) {
+        xchg_spi(0xFF); /* Dummy clock (force DO enabled) */
+    }
+    CS_HIGH();
+    sleep_ms(10);
+
     return _select();
+}
+
+uint pico_fatfs_get_clk_slow_freq(void)
+{
+    return _clk_slow_freq;
+}
+
+uint pico_fatfs_get_clk_fast_freq(void)
+{
+    return _clk_fast_freq;
 }