+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 * Add SPI PIO support for the case that pin assignment is not compliant with SPI function
+* Send 32 cycles of CS low in pico_fatfs_reboot_spi()
+### Changed
+* Update the benchmark test logs
 
 ## [fatfs-R0.15-1.0.2] - 2025-04-20
 ### Added